:recycle: Create custom hook for use in both return and send asset file
